requirements.txt
opencv-python==4.5.3.56
tensorflow-gpu==2.6.0
pandas==1.3.1
numpy==1.19.2
h5py==3.1.0
tables==3.6.1
joblib==1.0.1
imageio==2.9.0
openpyxl==3.0.7
flask==2.0.1
jinja2==3.0.1
git+https://github.com/tensorflow/docs
